- project re-structuring:   - moved sources     - src/* -> src/js     - test  -> src/test   - changed compilation/publishing tasks
