+++ .eslintrc
@@ -1,7 +1,5 @@
 parser: "babel-eslint"
 plugins: ["eslint-plugin-react"]
-ecmaFeatures:
-    jsx: true
 env:
     browser: true
     node: true
