+++ index.js
@@ -1,1 +1,1 @@
-module.exports = require('./lib/index');
+module.exports = require('./lib/js/index');
