+++ package.json
@@ -4,13 +4,16 @@
   "description": "React Tree View Component",
   "main": "index.js",
   "scripts": {
-    "prepublish": "npm run lib",
-    "lib": "npm run babel",
-    "babel": "rimraf lib && babel src/ -d lib/",
+    "css-compile": "node-sass --output-style expanded --source-map true --precision 6 src/scss/react-treee.scss lib/css/react-treee.css",
+    "scss-copy": "mkdir -p lib/scss && cp -r src/scss lib/",
+    "js-compile": "npm run babel",
+    "js-lint": "eslint src/js",
+    "prepublish": "npm-run-all clean-lib --parallel js-compile css-compile scss-copy",
+    "clean-lib": "rimraf lib",
+    "babel": "babel src/js -d lib/js",
     "test": "karma start karma.conf.js",
     "test-travis": "karma start --browsers Firefox --single-run",
-    "example": "webpack-dev-server --content-base ./example/ --config ./example/webpack.config.js",
-    "css-compile": "node-sass --output-style expanded --source-map true --precision 6 scss/react-treee.scss lib/css/react-treee.css"
+    "example": "npm run prepublish && webpack-dev-server --content-base ./example/ --config ./example/webpack.config.js"
   },
   "peerDependencies": {
     "react": "^15.5.4 || ^16.0.0",
@@ -63,6 +66,7 @@
     "mocha": "^3.4.2",
     "node-libs-browser": "^2.0.0",
     "node-sass": "^4.9.3",
+    "npm-run-all": "^4.1.3",
     "react": "^15.5.4",
     "react-dom": "^15.5.4",
     "react-hot-loader": "^1.3.0",
