+++ src/js/model/core-model.js
