+++ src/scss/_tree.scss
@@ -15,7 +15,7 @@
 .rt-link {
   cursor: pointer;
   position: relative;
-  padding: 0px 5px;
+  padding: 0 5px;
   display: block;
 }
