+++ src/test/model/tree-tests.js
@@ -1,4 +1,4 @@
-import Tree from '../../../src/model/tree';
+import Tree from '../../js/model/tree';
 
 describe('model/Tree', () => {
